chore(refactor): web/ and commonsjs/ layout
compatability with little-elements 1.1.+
