+++ .gitignore
@@ -19,3 +19,4 @@ node_modules/
 /bin/
 /site/
 /maps/
+/web/
