+++ karma.conf.js
@@ -15,14 +15,14 @@ module.exports = function(config) {
     // list of files / patterns to load in the browser
     // list of files / patterns to load in the browser
     files: [
-      'node_modules/@littleware/little-elements/lib/test/karmaAdapter.js',
+      'node_modules/@littleware/little-elements/web/lib/test/karmaAdapter.js',
       { pattern: 'node_modules/@littleware/**/*.js', type: 'module', included: false },
-      { pattern: 'lib/511/**/*.js', type: 'module', included: false },
-      { pattern: 'lib/jwt/**/*.js', type: 'module', included: false },
-      { pattern: 'lib/headerSimple/**/*.js', type: 'module', included: false },
+      { pattern: 'web/lib/511/**/*.js', type: 'module', included: false },
+      { pattern: 'web/lib/jwt/**/*.js', type: 'module', included: false },
+      { pattern: 'web/lib/headerSimple/**/*.js', type: 'module', included: false },
       { pattern: 'node_modules/lit-html/*.js', type: 'module', included: false },
       { pattern: 'node_modules/font-awesome/**/*', included: false },
-      { pattern: 'lib/testMain.js', type: 'module', included: true }
+      { pattern: 'web/lib/testMain.js', type: 'module', included: true }
     ],
     
     // list of files to exclude
@@ -35,7 +35,7 @@ module.exports = function(config) {
     },
     proxies: {
       '/lit-html': '/base/node_modules/lit-html',
-      '/little-elements': '/base/node_modules/@littleware/little-elements',
+      '/@littleware/little-elements': '/base/node_modules/@littleware/little-elements',
       '/modules/font-awesome': '/base/node_modules/font-awesome'
     },
