+++ package.json
@@ -1,10 +1,10 @@
 {
   "name": "@littleware/little-apps",
-  "version": "1.2.0",
+  "version": "1.3.0",
   "description": "littleware web clients",
   "main": "build/server.js",
   "dependencies": {
-    "@littleware/little-elements": "^1.0.3",
+    "@littleware/little-elements": "^1.1.0",
     "@webcomponents/webcomponentsjs": "^2.1.3",
     "express": "^4.14.0",
     "font-awesome": "^4.7.0",
