+++ src/@littleware/little-apps/lib/511/511.ts
@@ -1,6 +1,6 @@
-import { html, render } from '../../../../lit-html/lit-html.js';
-import '../../../little-elements/lib/arrivalPie/arrivalPie.js';
-import {singleton as styleHelper} from '../../../little-elements/lib/styleGuide/styleGuide.js';
+import { html, render } from '../../../../../lit-html/lit-html.js';
+import '../../../../../@littleware/little-elements/web/lib/arrivalPie/arrivalPie.js';
+import {singleton as styleHelper} from '../../../../../@littleware/little-elements/web/lib/styleGuide/styleGuide.js';
 import {css} from './511.css.js';
