+++ src/@littleware/little-apps/lib/511/index.html
@@ -1,10 +1,10 @@
-{% extends "little-elements/lib/styleGuide/shell/basicShell.html.njk" %}
+{% extends "little-elements/web/lib/styleGuide/shell/basicShell.html.njk" %}
 
 {% block head %}
 <link rel="stylesheet" type="text/css" href="../../../../jasmine-core/lib/jasmine-core/jasmine.css" />
 <script src="../../../../jasmine-core/lib/jasmine-core/jasmine.js"></script>
 <script src="../../../../jasmine-core/lib/jasmine-core/jasmine-html.js"></script>
-<script src="../../../little-elements/lib/test/littleJasmineBoot.js"></script>
+<script src="../../../little-elements/web/lib/test/littleJasmineBoot.js"></script>
 {% endblock %}
 
 {% block content %}
