+++ src/@littleware/little-apps/lib/headerSimple/index.html
@@ -1,4 +1,4 @@
-{% extends "little-elements/lib/styleGuide/shell/basicShell.html.njk" %}
+{% extends "little-elements/web/lib/styleGuide/shell/basicShell.html.njk" %}
 
 {% block head %}
 <title>headerView Test</title>
@@ -6,7 +6,7 @@
 <link rel="stylesheet" type="text/css" href="../../../../jasmine-core/lib/jasmine-core/jasmine.css" />
 <script src="../../../../jasmine-core/lib/jasmine-core/jasmine.js"></script>
 <script src="../../../../jasmine-core/lib/jasmine-core/jasmine-html.js"></script>
-<script src="../../../little-elements/lib/test/littleJasmineBoot.js"></script>
+<script src="../../../little-elements/web/lib/test/littleJasmineBoot.js"></script>
 {% endblock %}
 
 {% block content %}
