+++ src/@littleware/little-apps/lib/headerSimple/spec/main.ts
@@ -1,4 +1,4 @@
 import './headerSimpleSpec.js';
-import {startTest} from '../../../../little-elements/lib/test/util.js';
+import {startTest} from '../../../../../../@littleware/little-elements/web/lib/test/util.js';
 
 startTest();
