+++ src/@littleware/little-apps/lib/index.html
@@ -4,7 +4,7 @@
         <link rel="stylesheet" type="text/css" href="../../../jasmine-core/lib/jasmine-core/jasmine.css" />
         <script src="../../../jasmine-core/lib/jasmine-core/jasmine.js"></script>
         <script src="../../../jasmine-core/lib/jasmine-core/jasmine-html.js"></script>
-        <script src="../../little-elements/lib/test/littleJasmineBoot.js"></script>
+        <script src="../../little-elements/web/lib/test/littleJasmineBoot.js"></script>
 {% endblock %}
 
 {% block content %}
