+++ src/@littleware/little-apps/lib/jwt/jwt.css.ts
@@ -1,4 +1,4 @@
-import { html, render, TemplateResult } from '../../../../lit-html/lit-html.js';
+import { html, render, TemplateResult } from '../../../../../lit-html/lit-html.js';
 
 export const css = html`
 <style id="lw-jwt">
