+++ src/@littleware/little-apps/lib/jwt/spec/jwtSpec.ts
@@ -1,4 +1,4 @@
-import {getStage} from '../../../../little-elements/lib/test/util.js';
+import {getStage} from '../../../../../../@littleware/little-elements/web/lib/test/util.js';
 import {LittleJwt, parseSignedJwt} from '../jwt.js';
