+++ src/@littleware/little-apps/lib/jwt/spec/main.ts
@@ -1,5 +1,5 @@
 import './jwtSpec.js';
-import {startTest} from '../../../../little-elements/lib/test/util.js';
+import {startTest} from '../../../../../../@littleware/little-elements/web/lib/test/util.js';
 import '../../headerSimple/headerSimple.js';
 
 startTest();
