+++ src/@littleware/little-apps/lib/testMain.ts
@@ -1,6 +1,6 @@
 import './headerSimple/spec/headerSimpleSpec.js';
 import './511/spec/511Spec.js';
 import './jwt/spec/jwtSpec.js';
-import { startTest } from '../../little-elements/lib/test/util.js';
+import { startTest } from '../../../../@littleware/little-elements/web/lib/test/util.js';
 
 startTest();
