+++ src/@littleware/little-apps/site/511/index.html
@@ -1,4 +1,4 @@
-﻿{% extends "little-elements/lib/styleGuide/shell/basicShell.html.njk" %}
+﻿{% extends "little-elements/web/lib/styleGuide/shell/basicShell.html.njk" %}
 
 {% block head %}
   <title>5-1-1</title>
@@ -23,5 +23,5 @@
         </div>
     </div>
         
-    <script src="{{ jsroot }}/@littleware/little-apps/lib/511/main.js" type="module"></script>
+    <script src="{{ jsroot }}/@littleware/little-apps/web/lib/511/main.js" type="module"></script>
 {% endblock %}
