+++ src/@littleware/little-apps/site/auth/index.html
@@ -1,4 +1,4 @@
-{% extends "little-elements/lib/styleGuide/shell/basicShell.html.njk" %}
+{% extends "little-elements/web/lib/styleGuide/shell/basicShell.html.njk" %}
 
     {% block head %}
         <title>5-1-1</title>
