+++ src/@littleware/little-apps/site/index.html
@@ -1,4 +1,4 @@
-﻿{% extends "little-elements/lib/styleGuide/shell/basicShell.html.njk" %}
+﻿{% extends "little-elements/web/lib/styleGuide/shell/basicShell.html.njk" %}
 
 {% block head %}
 <title>frickjack.com</title>
@@ -54,17 +54,17 @@
 
                 <div class="pure-menu">
                     <ul class="pure-menu-list" id="appmenu">
-                        <li class="pure-menu-item"><a href="{{ jsroot }}/@littleware/little-elements/lib/test/index.html" class="pure-menu-link" title="test">test support</a></li>
-                        <li class="pure-menu-item"><a href="{{ jsroot }}/@littleware/little-elements/lib/arrivalPie/index.html" class="pure-menu-link" title="arrival-pie custom element">lw-arrival-pie custom element</a></li>
-                        <li class="pure-menu-item"><a href="{{ jsroot }}/@littleware/little-apps/lib/511/index.html" class="pure-menu-link" title="header-simple custom element">lw-511 custom element</a></li>
-                        <li class="pure-menu-item"><a href="{{ jsroot }}/@littleware/little-apps/lib/headerSimple/index.html" class="pure-menu-link" title="header-simple custom element">lw-header-simple custom element</a></li>
-                        <li class="pure-menu-item"><a href="{{ jsroot }}/@littleware/little-apps/lib/jwt/index.html" class="pure-menu-link" title="header-simple custom element">lw-jwt custom element</a></li>
-                        <li class="pure-menu-item"><a href="{{ jsroot }}/@littleware/little-elements/lib/styleGuide/index.html" class="pure-menu-link" title="littleware style guide">style guide</a></li>
+                        <li class="pure-menu-item"><a href="{{ jsroot }}/@littleware/little-elements/web/lib/test/index.html" class="pure-menu-link" title="test">test support</a></li>
+                        <li class="pure-menu-item"><a href="{{ jsroot }}/@littleware/little-elements/web/lib/arrivalPie/index.html" class="pure-menu-link" title="arrival-pie custom element">lw-arrival-pie custom element</a></li>
+                        <li class="pure-menu-item"><a href="{{ jsroot }}/@littleware/little-apps/web/lib/511/index.html" class="pure-menu-link" title="header-simple custom element">lw-511 custom element</a></li>
+                        <li class="pure-menu-item"><a href="{{ jsroot }}/@littleware/little-apps/web/lib/headerSimple/index.html" class="pure-menu-link" title="header-simple custom element">lw-header-simple custom element</a></li>
+                        <li class="pure-menu-item"><a href="{{ jsroot }}/@littleware/little-apps/web/lib/jwt/index.html" class="pure-menu-link" title="header-simple custom element">lw-jwt custom element</a></li>
+                        <li class="pure-menu-item"><a href="{{ jsroot }}/@littleware/little-elements/web/lib/styleGuide/index.html" class="pure-menu-link" title="littleware style guide">style guide</a></li>
                     </ul>
                 </div>
             </section>
         </div>
     </div>
 
-    <script src="{{ jsroot }}/@littleware/little-apps/lib/indexMain.js" type="module"></script>
+    <script src="{{ jsroot }}/@littleware/little-apps/web/lib/indexMain.js" type="module"></script>
 {% endblock %}
