+++ src/@littleware/little-apps/site/jwt/index.html
@@ -1,4 +1,4 @@
-﻿{% extends "little-elements/lib/styleGuide/shell/basicShell.html.njk" %}
+﻿{% extends "little-elements/web/lib/styleGuide/shell/basicShell.html.njk" %}
 
 {% block head %}
   <title>jwt</title>
@@ -21,5 +21,5 @@
         </div>
     </div>
         
-    <script src="{{ jsroot }}/@littleware/little-apps/lib/jwt/main.js" type="module"></script>
+    <script src="{{ jsroot }}/@littleware/little-apps/web/lib/jwt/main.js" type="module"></script>
 {% endblock %}
