+++ tsconfig.json
@@ -1,13 +1,12 @@
 {
     "compilerOptions": {
         //noImplicitAny: true,
-        "target": "es6",
+        "target": "ESNEXT",
         "sourceMap": true,
         "declaration": true,
         "sourceRoot": "src",
-        "baseUrl": "src",
         "rootDirs": [
-            "src",
+            ".",
             "node_modules"
         ],
         "traceResolution": true
